pyproject.toml
[tool.poetry]
name = "exercise-app"
version = "0.1.0"
description = ""
authors = ["Jorard <jorard@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.1"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.17"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
